refactor(dir): rename some files using kebab-case
