+++ src/errors/error-handler.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from 'express';
 import { StatusCodes, getReasonPhrase } from 'http-status-codes';
 import { ErrorResponse } from '../types';
-import { AppError } from './appErrors';
+import { AppError } from './app-errors';
 
 export const errorHandler = (err: AppError, _req: Request, res: Response, _next: NextFunction) => {
   const errorResponse: ErrorResponse = {
