+++ src/errors/index.ts
@@ -4,6 +4,9 @@ export {
   MethodNotAllowedError,
   NotFoundError,
   ValidationError,
+} from './app-errors';
+export { errorHandler } from './error-handler';
+export { prismaErrorHandler } from './prisma-error-handler';
 } from './appErrors';
 export { errorHandler } from './errorHandler';
 export { prismaErrorHandler } from './prismaErrorHandler';
